TL: merging generic Collocation class and CollBase in core
