+++ pySDC/core/Collocation.py
@@ -1,17 +1,61 @@
 import logging
 
 import numpy as np
+import scipy.interpolate as intpl
 
+from pySDC.core.Nodes import NodesGenerator
 from pySDC.core.Errors import CollocationError
 from pySDC.core import LagrangeApproximation
 
 
 class CollBase(object):
     """
-    Abstract class for collocation
+    Generic collocation class, that contains everything to do integration over
+    intervals and between nodes.
+    It can be used to produce many kind of quadrature nodes from various
+    distribution (awesome!).
 
-    Derived classes will contain everything to do integration over intervals and between nodes, they only need to
-    provide the set of nodes, the rest is done here (awesome!)
+    It is based on the two main parameters that define the nodes :
+
+    - node_type : the node distribution used for the collocation method
+    - quad_type : the type of quadrature used (inclusion of not of boundary)
+
+    Current implementation provides the following available parameter values
+    for node_type :
+
+    - EQUID : equidistant node distribution
+    - LEGENDRE : distribution from Legendre polynomials
+    - CHEBY-{1,2,3,4} : distribution from Chebyshev polynomials of a given kind
+
+    The type of quadrature cann be GAUSS (only inner nodes), RADAU-LEFT
+    (inclusion of the left boundary), RADAU-RIGHT (inclusion of the right
+    boundary) and LOBATTO (inclusion of left and right boundary).
+
+    Furthermore, the ``useSpline`` option can be activated to eventually use
+    spline interpolation when computing the weights.
+
+    Here is the equivalency table with the original classes implemented in
+    pySDC :
+
+    +-------------------------+-----------+-------------+-----------+
+    | Original Class          | node_type | quad_type   | useSpline |
+    +=========================+===========+=============+===========+
+    | Equidistant             | EQUID     | LOBATTO     | False     |
+    +-------------------------+-----------+-------------+-----------+
+    | EquidistantInner        | EQUID     | GAUSS       | False     |
+    +-------------------------+-----------+-------------+-----------+
+    | EquidistantNoLeft       | EQUID     | RADAU-RIGHT | False     |
+    +-------------------------+-----------+-------------+-----------+
+    | EquidistantSpline_Right | EQUID     | RADAU-RIGHT | True      |
+    +-------------------------+-----------+-------------+-----------+
+    | CollGaussLegendre       | LEGENDRE  | GAUSS       | False     |
+    +-------------------------+-----------+-------------+-----------+
+    | CollGaussLobatto        | LEGENDRE  | LOBATTO     | False     |
+    +-------------------------+-----------+-------------+-----------+
+    | CollGaussRadau_Left     | LEGENDRE  | RADAU-LEFT  | False     |
+    +-------------------------+-----------+-------------+-----------+
+    | CollGaussRadau_Right    | LEGENDRE  | RADAU-RIGHT | False     |
+    +-------------------------+-----------+-------------+-----------+
 
     Attributes:
         num_nodes (int): number of collocation nodes
@@ -26,7 +70,8 @@ class CollBase(object):
         left_is_node (bool): flag to indicate whether left point is collocation node
     """
 
-    def __init__(self, num_nodes, tleft=0, tright=1):
+    def __init__(self, num_nodes, tleft=0, tright=1,
+                 node_type='LEGENDRE', quad_type='LOBATTO', useSpline=False):
         """
         Initialization routine for an collocation object
 
@@ -48,14 +93,29 @@ class CollBase(object):
         self.tleft = tleft
         self.tright = tright
 
-        # Dummy values for the rest
-        self.nodes = None
-        self.weights = None
-        self.Qmat = None
-        self.Smat = None
-        self.delta_m = None
-        self.right_is_node = None
-        self.left_is_node = None
+        # Instanciate attributes
+        self.nodeGenerator = NodesGenerator(node_type, quad_type)
+        if useSpline:
+            self._getWeights = self._getWeights_spline
+            # We need: 1<=order<=5 and order < num_nodes
+            self.order = min(num_nodes - 1, 3)
+        elif node_type == 'EQUID':
+            self.order = num_nodes
+        else:
+            if quad_type == 'GAUSS':
+                self.order = 2 * num_nodes
+            elif quad_type.startswith('RADAU'):
+                self.order = 2 * num_nodes - 1
+            elif quad_type == 'LOBATTO':
+                self.order = 2 * num_nodes - 2
+
+        self.nodes = self._getNodes
+        self.weights = self._getWeights(tleft, tright)
+        self.Qmat = self._gen_Qmatrix_spline if useSpline else self._gen_Qmatrix
+        self.Smat = self._gen_Smatrix
+        self.delta_m = self._gen_deltas
+        self.left_is_node = quad_type in ['LOBATTO', 'RADAU-LEFT']
+        self.right_is_node = quad_type in ['LOBATTO', 'RADAU-RIGHT']
 
     @staticmethod
     def evaluate(weights, data):
@@ -98,11 +158,26 @@ class CollBase(object):
 
         return np.ravel(weights)
 
+    @property
     def _getNodes(self):
         """
-        Dummy method for generating the collocation nodes.
+        Computes nodes using an internal NodesGenerator object
+
+        Returns:
+            np.ndarray: array of Gauss-Legendre nodes
         """
-        raise NotImplementedError('ERROR: collocation has to implement _getNodes(self)')
+        # Generate normalized nodes in [-1, 1]
+        nodes = self.nodeGenerator.getNodes(self.num_nodes)
+
+        # Scale nodes to [tleft, tright]
+        a = self.tleft
+        b = self.tright
+        nodes += 1
+        nodes /= 2
+        nodes *= b - a
+        nodes += a
+
+        return nodes
 
     @property
     def _gen_Qmatrix(self):
@@ -163,3 +238,46 @@ class CollBase(object):
             delta[m] = self.nodes[m] - self.nodes[m - 1]
 
         return delta
+
+    def _getWeights_spline(self, a, b):
+        """
+        Computes weights using spline interpolation instead of Gaussian
+        quadrature
+
+        Args:
+            a (float): left interval boundary
+            b (float): right interval boundary
+
+        Returns:
+            np.ndarray: weights of the collocation formula given by the nodes
+        """
+
+        # get the defining tck's for each spline basis function
+        circ_one = np.zeros(self.num_nodes)
+        circ_one[0] = 1.0
+        tcks = []
+        for i in range(self.num_nodes):
+            tcks.append(
+                intpl.splrep(self.nodes, np.roll(circ_one, i), xb=self.tleft, xe=self.tright, k=self.order, s=0.0))
+
+        weights = np.zeros(self.num_nodes)
+        for i in range(self.num_nodes):
+            weights[i] = intpl.splint(a, b, tcks[i])
+
+        return weights
+
+    @property
+    def _gen_Qmatrix_spline(self):
+        """
+        Compute tleft-to-node integration matrix for later use in collocation formulation
+        Returns:
+            numpy.ndarray: matrix containing the weights for tleft to node
+        """
+        M = self.num_nodes
+        Q = np.zeros([M + 1, M + 1])
+
+        # for all nodes, get weights for the interval [tleft,node]
+        for m in np.arange(M):
+            Q[m + 1, 1:] = self._getWeights(self.tleft, self.nodes[m])
+
+        return Q
