+++ pySDC/core/__init__.py
@@ -1,5 +1,6 @@
 from pySDC.core.Nodes import NodesGenerator
 from pySDC.core.Lagrange import LagrangeApproximation
+from pySDC.core.Collocation import CollBase
 
 __author__ = 'robert'
-__all__ = ['NodesGenerator', 'LagrangeApproximation']
+__all__ = ['NodesGenerator', 'LagrangeApproximation', 'CollBase']
