+++ pySDC/implementations/collocation_classes/equidistant_spline_right.py
@@ -1,7 +1,7 @@
-from pySDC.implementations.collocations import Collocation
+from pySDC.core import CollBase
 
 
-class EquidistantSpline_Right(Collocation):
+class EquidistantSpline_Right(CollBase):
 
     def __init__(self, num_nodes, tleft, tright):
         super(EquidistantSpline_Right, self).__init__(
