+++ pySDC/tests/test_collocation.py
@@ -4,7 +4,6 @@ import numpy as np
 from pySDC.core.Collocation import CollBase
 from pySDC.tests.test_helpers import get_derived_from_in_package
 
-from pySDC.implementations.collocations import Collocation
 from pySDC.implementations.collocation_classes.equidistant import \
     Equidistant
 from pySDC.implementations.collocation_classes.equidistant_inner import \
@@ -50,7 +49,7 @@ def testEquivalencies():
     # Compare each original class with their equivalent generic implementation
     for params, CollClass in EQUIV.items():
         cOrig = CollClass(M, tLeft, tRight)
-        cNew = Collocation(M, tLeft, tRight, *params)
+        cNew = CollBase(M, tLeft, tRight, *params)
         for attr in lAttrVect:
             assert norm(getattr(cOrig, attr)-getattr(cNew, attr)) < tol
         for attr in lAttrScalar:
